add addOrder

+++ cnpm/src/main/java/com/doan/cnpm/entity/Include.java
@@ -33,5 +33,5 @@ public class Include {
     @JoinColumn(name = "productID", referencedColumnName = "productID") // Khóa ngoại cho Product
     private Product product;
 
-    private int quantity;
+    private Long quantity;
 }
